Re-structuring of modules
- Moved data processing into process sub-module
- Added data processing for MPANT .mpa files and MCDWIN .887 files,
adopted from D. Atanasov dEval
- Moved Peaks and softCool into utilities sub-module
- Added readme.md for submodules
