+++ lukas-eval/bin/process.py
@@ -1,315 +1,393 @@
-import ROOT
-from ROOT import RooFit
-from ROOT import RooRealVar, RooArgSet, RooArgList, RooDataHist
-from ROOT import RooGenericPdf, RooUniform, RooGaussian, RooGaussModel, RooDecay, RooFormulaVar
-from ROOT import RooAddPdf, RooMCStudy
-from ROOT import Math as RootMath 
-import pandas as pd
+# -*- coding: utf-8 -*-
+"""
+Created on Mon 11 January 2022
+Modified by Lukas.Nies@cern.ch on 21/10/2021
+@author: Lukas Nies
+@contact: Lukas.Nies@cern.ch
+@license: /
+"""
+
+import mmap
+import os
+import sys
 import numpy as np
-import matplotlib.pyplot as plt
-import matplotlib as mpl
-import scipy as sc
-from scipy.signal import find_peaks
-from scipy.optimize import least_squares
-from scipy import stats
-import math 
-import time
+import pandas as pd
+from io import StringIO
+from configparser import ConfigParser
 
-class Peaks:
-	""" Wrapper class for finding peaks in an MR-ToF MS spectrum"""
-	# df_file: dataframe containing the converted .lst content 
-	#
-	#
-	def __init__(self, df_file):
-		self.file = df_file
-		self.n_peaks = 0
-		self.bins = 500
-		self.peak_threshold = 0.002
-		self.peak_min_distance = 5
-		self.peak_min_height = 10
-		self.peak_width_inbins = (3,100)
-		self.peak_prominence = None
-		self.peak_wlen = None
+class CustomParser(ConfigParser):
+    """
+    Original from https://gitlab.cern.ch/datanaso/dEval/-/blob/master/etc/utils.py
+    ConfigParser with a custom dictionary conversion method.
+    """
+    def as_dict(self):
+        d = dict(self._sections)
+        for k in d:
+            d[k] = dict(self._defaults, **d[k])
+            d[k].pop('__name__', None)
+        return d
 
-	def get_binning(self, bins=10):
-		"""
-		Adapts binning to multiples of 0.8ns, assuming that 0.8ns was used to take data (to common case)
-		"""
-		# Get min and max tof from data frame
-		minn = self.file.tof.min()
-		maxx = self.file.tof.max()
-		return round((maxx-minn)/0.8/bins)
-	
-	def find_peaks(self, bins=10, peak_threshold = 0.002, peak_min_distance = 5, peak_min_height = 10, peak_width_inbins = (3,100), 
-				   peak_prominence = None, peak_wlen = None):
-		"""  """
-		#
-		self.pos = []
-		self.std = []
-		self.left_bases = []
-		self.right_bases = []
-		self.bins = bins
-		# faster binning for projections than histograms -> necessary in order to automatically find peaks
-		x_proj_for_pfind = self.file.tof.value_counts(bins=self.get_binning(self.bins)).sort_index()
-		y_proj_for_pfind = self.file.sweep.value_counts(bins=self.get_binning(self.bins)).sort_index()
-		# Do the peak finding
-		self.x_proj_peaks, self.peaks_info = sc.signal.find_peaks(x_proj_for_pfind, threshold=peak_threshold,
-											 distance=peak_min_distance,
-											 height=peak_min_height,
-											 width=peak_width_inbins,
-											 prominence=peak_prominence,
-											 wlen=peak_wlen)
-		# Calculate some additional meta data for the found peaks
-		self.highest_peak = self.peaks_info['peak_heights'].argmax()
-		self.n_peaks = len(self.x_proj_peaks)        
-		# variables to store earliest left_base and lastest right_base for constraining plot range
-		self.earliest_left_base = 1e15
-		self.earliest_peak_idx = 1e15
-		self.latest_right_base = 0
-		self.latest_peak_idx = 0
-		for i in range(self.n_peaks):
-			# get the peak bases ranges from the peak finder
-			left = x_proj_for_pfind.index.mid[self.peaks_info['left_bases'][i]]
-			right = x_proj_for_pfind.index.mid[self.peaks_info['right_bases'][i]]
-			self.left_bases.append(left)
-			self.right_bases.append(right)
-			#calculate the median (more accurate due to asym. tails) of the data in the peak ranges
-			#better value for the actual peak center than the simple highest point in peak
-			peak_pos = self.file.tof[(self.file.tof < right) &
-								  (self.file.tof > left)].median()
-			peak_std = self.file.tof[(self.file.tof < right) &
-								  (self.file.tof > left)].std()
-			# estimate the mean and sigma from a Gaussian with exponential tail (accounting for asym. peaks)
-			try: 
-				peak_fit = stats.exponnorm.fit(file_df.tof[(file_df.tof < peak_pos+peak_std) &
-														   (file_df.tof > peak_pos-peak_std)],
-											   loc=peak_pos)
-				# update peak position if a fit was possible
-				peak_pos = peak_fit[1]
-			except:
-				pass
-			self.pos.append(peak_pos)
-			self.std.append(peak_std)
-			# assign earliest and latest bases
-			if (left < self.earliest_left_base and not math.isnan(left) and not math.isnan(peak_std)): 
-				self.earliest_left_base = left 
-				self.earliest_peak_idx = i 
-			if (right > self.latest_right_base and not math.isnan(right) and not math.isnan(peak_std)):
-				self.latest_right_base = right
-				self.latest_peak_idx = i 
-				
-	def plot(self, bins = 10, focus=False, log=False, silent = False, save = False, path_to_file = "peaks"):
-		#
-		plt.rcParams["figure.figsize"] = (10,6)
-		#
-		if self.n_peaks == 0:
-			print("Not peaks, no plots :)")
-			return 0 
-		#            
-		xdata = self.file.tof
-		n, xe = np.histogram(xdata, bins=self.get_binning(bins))
-		cx = 0.5 * (xe[1:] + xe[:-1])
-		dx = np.diff(xe)
-		plt.errorbar(cx, n, n ** 0.5, fmt="ok", zorder=1)
-		plt.plot(xdata, np.zeros_like(xdata)-5, "|", alpha=0.1, label = "ToF Data", zorder = 3)
-		if log:
-			plt.yscale('log')
-		for i in range(self.n_peaks):
-			plt.axvline(self.pos[i], c='r', linewidth=1, zorder=3)
-		
-		xm = np.linspace(xe[0], xe[-1], num=1000)
-		plt.legend();
-		# plt.xlim(peaks.pos[0]-300, peaks.pos[0]+300)
-		# Zoom in on found peaks
-		if focus==False:
-			plt.xlim(self.earliest_left_base-200, self.latest_right_base+200)
-		else:
-			plt.xlim(self.pos[focus]-300, self.pos[focus]+300)
-		#
-		if not silent: 
-			plt.show()
-			# plt.clf()
+class MCS6Lst():
+    '''
+    Process each list file contained in array of lst_files
+    Param:
+        lst_files: array of list files
+    '''
+    """
+    Created on Mon 23 April 2018
+    Modified and adapted to Python 3 on Wed 17 July 2019
+    Modified by Lukas.Nies@cern.ch on 21/10/2021
+    @author: Maxime Mougeot
+    @author: Jonas Karthein
+    @contact: maxime.mougeot@cern.ch
+    @contact: jonas.karthein@cern.ch
+    @license: MIT license
+    """
+    def __init__(self):
+        """
+        Initialize the conversion dataframe and some other varaibles
+        """
+        self.lst_files = []
+        #-------------------------Create a dataframe containing the conversion table from the MCS6A manual--------------------------------------#
+        Time_Patch_Value = ['0', '5', '1', '1a', '2a', '22', '32','2','5b','Db','f3','43','c3','3']
+        conversion_dict = {'Data_Length' : pd.Series([2,4,4,6,6,6,6,6,8,8,8,8,8,8], index=Time_Patch_Value),
+        'Data_Lost_Bit' : pd.Series([np.nan ,np.nan ,np.nan ,np.nan ,np.nan ,np.nan ,47,np.nan ,63,np.nan ,47,63,np.nan ,63], index=Time_Patch_Value),
+        'Tag_Bits' : pd.Series([(np.nan,np.nan) ,(np.nan,np.nan) ,(np.nan,np.nan) ,(np.nan,np.nan) ,(40,47),(40,47),(np.nan,np.nan), (np.nan,np.nan), (48,62),(48,63),(48,63),(48,62),(48,63),(58,62)], index=Time_Patch_Value),
+        'Sweep_Counter': pd.Series([(np.nan,np.nan),(24,31),(np.nan,np.nan),(32,47),(32,39),(np.nan,np.nan),(40,46),(np.nan,np.nan),(32,47),(32,47),(40,46),(np.nan,np.nan),(np.nan,np.nan),(np.nan,np.nan)], index=Time_Patch_Value),
+        'Time_Bits': pd.Series([12,20,28,28,28,36,36,44,28,28,36,44,44,54], index=Time_Patch_Value),
+        'Max_Sweep_Length': pd.Series([0.0000004096,0.000105,0.027,0.027,0.027,6.872,6.872,1759.2,0.027,0.027,6.872,1759.2,1759.2,1801440], index=Time_Patch_Value)}
+        self.conversion_df = pd.DataFrame(conversion_dict)
+        self.df_dict = {}
 
-		if save:
-			plt.savefig(path_to_file+".pdf", dpi=300)
-			plt.clf()
-		
-	def plot2d(self, bins=500, focus=-1, log=False):
-		"""
-		
-		"""
-		plt.rcParams["figure.figsize"] = (5,15)
-		tof = self.file.tof
-		sweep = self.file.sweep
-		plt.plot(tof, sweep, 'o', alpha=0.05, ms=2, label='unbinned data')
-		for i in range(self.n_peaks):
-			plt.axvline(self.pos[i], c='r', linewidth=1, zorder=3)
-		if focus==False:
-			plt.xlim(self.earliest_left_base-200, self.latest_right_base+200)
-		else:
-			plt.xlim(self.pos[focus]-300, self.pos[focus]+300)
-		plt.xlabel("Time-of-Flight [ns]")
-		plt.ylabel("Rolling sweep number")
-		plt.show()
+    def convert_bytes(self,bytearray,nb_bits,data_lost_bit,tag_bits,sweep_counter,time_bits,verbose=0):
+        '''
+        Perform the actual conversion of a single event from binary to integer numbers
+        See pages  5-19 and 5-20 of FastCom MCS6A for more information on the bits meaning
+        Param:
+            bytearray : an array of nb_bits/8 bytes encapsulating the data of a single MCS6A stop event
+            nb_bits : total number of bits on which the data is encoded
+            data_lost_bit : Data lost bit. Indicates if the fifo was fullself. 1 bit index
+            tag_bits : Tag bits. Tag info of a single stop event (see manual). Tuple containing the bit indexes.
+            sweep_counter: Sweep number of a single stop event. Tuple containing the bit indexes.
+            time_bits : Number of bits encoding the time of flight a single stop eventself.
+            The tof seems to be given in an unbinned format with 100ps resolution (to be confirmed).
+        Return:
+            Decoded tof, sweep, channel, edge, tag, fifo
+        '''
+        bit_word = ''
 
-class softCool(Peaks):
-	""" Class for performing software cooling on 2D MR-ToF MS Data"""
-	# df_file: dataframe containing the converted .lst content 
-	# Inherits functionality from the peak finder 
-	#
-	def __init__(self, df_file, peaks = False):
-		"""
-		Initialize by finding peaks in the ToF
-		""" 
-		self.file = df_file
-		self.coolfile = df_file.copy(deep=True) # copy for storing the cooled spectrum
-		if peaks == False:
-			self.peaks=Peaks(self.file)
-			self.peaks.find_peaks(bins=200,
-							peak_threshold=2,
-							peak_min_distance=5,
-							peak_min_height=100,
-							peak_width_inbins=(0.5, 100),
-							peak_prominence=1,
-							peak_wlen=50)
-		else:
-			self.peaks = peaks
-		self.selected_peak = -1
-		self.corr_factors = []
-		self.chunk_size = 10
-		self.post_cool = False
-		
-	def select_peak(self, peak_nb):
-		"""
-		
-		"""
-		self.selected_peak = peak_nb
-	
-	def calc_corr_factors(self, df, tof_cut, chunk_size=10, method="mean"):
-		"""
-		Function for calculating correction factors
-		"""
-		df_cut = df[(df.tof > tof_cut[0]) & (df.tof < tof_cut[1])]
-		self.chunk_size = chunk_size
-		#
-		if method=="mean":
-			self.corr_factors = [
-				 np.mean(sublist.tof) 
-				 for sublist  
-				 in 
-					[
-						 df_cut[(df_cut.sweep >= i) & (df_cut.sweep < i+self.chunk_size)]
-						 for i 
-						 in range(0,df.sweep.iloc[-1], self.chunk_size)
-					]
-			]
-		if method=="median":
-			self.corr_factors = [
-				 np.median(sublist.tof) 
-				 for sublist  
-				 in 
-					[
-						 df_cut[(df_cut.sweep >= i) & (df_cut.sweep < i+self.chunk_size)]
-						 for i 
-						 in range(0,df.sweep.iloc[-1], self.chunk_size)
-					]
-			]
-		if method=="average":
-			self.corr_factors = [
-				 np.average(sublist.tof) 
-				 for sublist  
-				 in 
-					[
-						 df_cut[(df_cut.sweep >= i) & (df_cut.sweep < i+self.chunk_size)]
-						 for i 
-						 in range(0,df.sweep.iloc[-1], self.chunk_size)
-					]
-			]
+        for bytes in reversed(bytearray):
+            bit_word += '{0:08b}'.format(ord(chr(bytes)))
 
-	def cool(self, method="mean", tof_cut_left=300, tof_cut_right=300, chunk_size=10, post_cool = False):
-		"""
-		Routine for performing the cooling
-		"""
-		# df to be cooled
-		self.post_cool = post_cool
-		if not self.post_cool:
-			df_to_cool = self.file
-		else:
-			df_to_cool = self.coolfile
-		#
-		self.chunk_size = chunk_size
-		tof_cut = [self.peaks.pos[self.selected_peak]-tof_cut_left, self.peaks.pos[self.selected_peak]+tof_cut_right]
-		self.calc_corr_factors(df_to_cool, tof_cut, self.chunk_size)
-		# print(f"Length correction factors: {len(self.corr_factors)}")
-		# print(f"Length chunk sizes: {len(range(0,df_to_cool.sweep.iloc[-1], self.chunk_size))}")
-		#
-		cooled_tofs = [
-			[
-				# print(cooled.corr_factors[j[1]], j[1])
-				row - self.corr_factors[j[1]] + self.corr_factors[0]
-				for row 
-				in j[0]
-			]
-			for j 
-			in
-			[
-				 [df_to_cool[(df_to_cool.sweep >= i) & (df_to_cool.sweep < i+self.chunk_size)].tof, int(i/self.chunk_size)]
-				 for i 
-				 in range(0,df_to_cool.sweep.iloc[-1], self.chunk_size)
-			 ]
-		]
-		# print(self.coolfile)
-		# print(self.coolfile.sweep.tolist())
-		# print(f"Length cooled tofs: {[len(tofs) for tofs in cooled_tofs]}, sum: {sum([len(tofs) for tofs in cooled_tofs])}")
-		# Flatten 2d array from previous list-comprehension
-		self.coolfile.tof = [
-			 item 
-			 for sublist 
-			 in 
-			 cooled_tofs for item in sublist
-		]
-		# self.coolfile.tof = np.hstack(cooled_tofs)
+        #convert data lost bit always first index in the reversed array (last in the manual)
+        if np.isnan(data_lost_bit):
+            fifo = np.nan
+            index_data_lost_bit = -1
+        else:
+            index_data_lost_bit = nb_bits-1-int(data_lost_bit)
+            fifo = int(bit_word[index_data_lost_bit],2)
 
-		# Drop empty sweeps
-		for idx in self.coolfile[self.coolfile.tof.isnull()].index:
-			self.coolfile = self.coolfile.drop(idx)
-		# Export the cooled df
-		return self.coolfile
-		
-	def plot2d(self, bins=500, focus=False, log=False):
-		"""
-		Plot the 2d histograms to compare the corrected and not corrected values
-		"""
-		fig, (ax0, ax1) = plt.subplots(1,2,sharey='row', figsize=(7,7))
-		tof = self.file.tof
-		sweep = self.file.sweep
-		# Plot unbinned and un-corrected data
-		ax0.plot(tof, sweep, 'o', alpha=0.05, ms=2, label='unbinned data')
-		# Plot correction factors
-		y_corr = range(0,self.file.sweep.iloc[-1], self.chunk_size)
-		x_corr = self.corr_factors
-		ax0.plot(x_corr, y_corr, c='r', linewidth=1, zorder=3)
-		# Plot corrected data
-		tof = self.coolfile.tof
-		sweep = self.coolfile.sweep
-		ax1.plot(tof, sweep, 'o', alpha=0.05, ms=2, label='unbinned data')
-		if self.post_cool:
-			y_corr = range(0,self.coolfile.sweep.iloc[-1], self.chunk_size)
-			x_corr = self.corr_factors
-			ax1.plot(x_corr, y_corr, c='r', linewidth=1, zorder=3)
-		#
-		# for i in range(self.n_peaks):
-		#     plt.axvline(self.pos[i], c='r', linewidth=1, zorder=3)
-		if focus==False:
-			ax0.set_xlim(self.peaks.earliest_left_base-300, self.peaks.latest_right_base+300)
-			ax1.set_xlim(self.peaks.earliest_left_base-300, self.peaks.latest_right_base+300)
-		else:
-			ax0.set_xlim(self.peaks.pos[focus]-300, self.peaks.pos[focus]+300)
-			ax1.set_xlim(self.peaks.pos[focus]-300, self.peaks.pos[focus]+300)
+        #convert tag bit
+        if np.isnan(tag_bits[0]):
+            tag = np.nan
+            index_high_tag = -1
+            index_low_tag = -1
+        else:
+            index_high_tag = nb_bits-1-int(tag_bits[1])
+            index_low_tag = nb_bits-1-int(tag_bits[0])
+            tag = int(bit_word[index_high_tag:index_low_tag+1],2)
 
-		plt.xlabel("Time-of-Flight [ns]")
-		plt.ylabel("Rolling sweep number")
+        #convert sweep number
+        if np.isnan(sweep_counter[0]):
+            sweep = np.nan
+            index_high_sweep = -1
+            index_low_sweep = -1
+        else:
+            index_high_sweep = nb_bits-1-int(sweep_counter[1])
+            index_low_sweep = nb_bits-1-int(sweep_counter[0])
+            sweep = int(bit_word[index_high_sweep:index_low_sweep+1],2)
 
-		plt.tight_layout()
-		plt.show()
+        if bit_word != "000000000000000000000000000000000000000000000000" and verbose>1:
+            print(f"bit_word: {bit_word}")
+            print(f"index_data_lost_bit: {fifo}")
+            print(f"index_high_tag: {index_high_tag}")
+            print(f"index_low_tag: {index_low_tag}")
+            print(f"tag: {tag}")
+            print(f"index_high_sweep: {index_high_sweep}")
+            print(f"index_low_sweep: {index_low_sweep}")
+            print(f"sweep: {sweep}")
 
+        #convert time of flight
+        index_high_tof = max(index_low_sweep,index_low_tag,index_data_lost_bit)+1
+        index_low_tof = index_high_tof+time_bits
+        tof = int(bit_word[index_high_tof:index_low_tof],2)
+
+        #these are always there no matter the format
+        channel = int(bit_word[index_low_tof+1:],2)
+        edge = int(bit_word[index_low_tof],2)
+
+        # if tof != 0:
+        #     print(tof, sweep-1, channel, edge, tag, fifo)
+
+        return tof, sweep-1, channel, edge, tag, fifo
+
+    def decode_binary(self,binary,time_patch, verbose = 0):
+        '''
+        Read the binary part of the file by chunks and decode each chunk according to the format
+        given in time_patch
+        The length of a chunk is given by the time_patch
+        Param : binary part of
+        Return: nunpy array containing the converted data : tof, sweep, channel, edge, tag, fifo
+        '''
+
+        #-----------extract data from the dataframe--------------------------#
+        data_length = int(self.conversion_df.loc[time_patch.decode('ascii'),'Data_Length'])
+        nb_bits = 8*data_length #convert nb of bytes into nb of bits
+        data_lost_bit = self.conversion_df.loc[time_patch.decode('ascii'),'Data_Lost_Bit']
+        tag_bits = self.conversion_df.loc[time_patch.decode('ascii'),'Tag_Bits']
+        sweep_counter = self.conversion_df.loc[time_patch.decode('ascii'),'Sweep_Counter']
+        time_bits = int(self.conversion_df.loc[time_patch.decode('ascii'),'Time_Bits'])
+        max_sweep_length = self.conversion_df.loc[time_patch.decode('ascii'),'Max_Sweep_Length']
+
+        steps = len(binary[binary.tell():])/data_length
+        first_it = True
+
+        if verbose>1:
+            print(f"Data length: {data_length}\nN bits: {nb_bits}\nData lost bit: {data_lost_bit}\n\
+                    tag bits: {tag_bits}\nsweep_counter: {sweep_counter}\ntime_bits: {time_bits}\n\
+                    max sweep length: {max_sweep_length}\nsteps: {steps}\n")
+        # !
+        # Introduce sweep_counter_overflow: in some cases, MCS6 seems to allocated only a small amount of bits for storing the sweep number.
+        # In time_patch=32 this is for example only 7 bits -> can count to 128 and then resets to 0. With the overflow counter we just
+        # count how many overflows happen and just at the necessary amount of sweeps to the overall sweep number
+        sweep_counter_overflow = 0
+        old_sweep = 0 # for detecting when overflow happens
+        # loop through all bytewords
+        for i in range(int(steps)):
+            if verbose>0:
+                if (i%(int(steps/10))==0):
+                    print(f"Step {i} of {steps}.")
+            byteword = binary.read(data_length)
+            tof, sweep, channel, edge, tag, fifo = self.convert_bytes(byteword,nb_bits,
+                data_lost_bit, tag_bits, sweep_counter, time_bits, verbose=verbose)
+            # Check whether overflow happened (for example old_sweep = 127, new sweep is 0)
+            # Only do for non-zero events:
+            if tof != 0: 
+                if verbose>1: print(f"old_sweep: {old_sweep}")
+                if old_sweep > sweep:
+                    sweep_counter_overflow += 1
+                if verbose>1: print(f"sweep_counter_overflow: {sweep_counter_overflow}")
+                old_sweep = sweep 
+                # Add overflow to the sweep number (in case sweep has 7bit int -> 2**7=128)
+                sweep += sweep_counter_overflow*(2**(sweep_counter[1]-sweep_counter[0]+1))
+                if verbose>1: print(f"sweep: {sweep}")
+            #
+            if channel != 0 :#means for real data
+                if first_it:
+                    converted_data = np.array([tof, sweep, channel, edge, tag, fifo])
+                    first_it = False
+                else :
+                    converted_data = np.vstack((converted_data, np.array([tof, sweep, channel, edge, tag, fifo])))
+        binary.close()
+        return converted_data
+
+    def get_time_patch_and_binary(self, listfile, verbose=False):
+        '''
+        Memory map the list file and isolate the time_patch and the binary part of the data
+        Param:
+            listfile : input list file
+        Return:
+            mapped_file : memore map of the input listfile
+            time_patch : string code indicating the format in which the data are written (see manual)
+        '''
+        mapped_file = mmap.mmap(listfile.fileno(), 0, access=mmap.ACCESS_READ)
+        search_dict = {'section' : '[DATA]' , 'list_file_type' : 'time_patch'}
+        #-----------------set file index to time patch code -----------------#
+        pos_type_from = mapped_file.find(search_dict['list_file_type'].encode('ascii'))+len(search_dict['list_file_type'])+1
+        mapped_file.seek(pos_type_from)
+        time_patch = mapped_file.readline().strip('\r\n'.encode('ascii'))
+        #-----------set file index to beginning of DATA-----------------------------------#
+        pos_data_from = mapped_file.find(search_dict['section'].encode('ascii'))
+        mapped_file.seek(pos_data_from)
+        #---readline and there no matter what the file index should point to the beginning of the binary data
+        mapped_file.readline()
+
+        if verbose>1:
+            print(f"pos_type_from: {pos_type_from}\npos_data_from: {pos_data_from}\ntime_patch: {time_patch}")
+
+        return mapped_file, time_patch
+
+    def process(self,file_array,to_csv = False, verbose=0):
+        """
+        Perform the processing of the lst_files 
+        """
+        full_info = False   # for regular application the channel, edge, tag and fifo info are constant so they don't have to be saved. In that case keep full_info = False
+        self.lst_files = file_array
+        for filename in self.lst_files:
+            with open(filename,'rb') as listfile:
+
+                binary, time_patch = self.get_time_patch_and_binary(listfile, verbose=verbose)
+
+                if full_info:
+                    converted_data = self.decode_binary(binary,time_patch,verbose=verbose)    # np.array with tof, sweep, channel, edge, tag, fifo
+                    header_res ='tof,sweep,channel,edge,tag,fifo'
+                    if to_csv:
+                        np.savetxt('{}/{}.csv'.format(os.path.split(filename)[0],os.path.splitext(os.path.basename(filename))[0]),converted_data,
+                    fmt = '%i,%i,%i,%i,%f,%f', header = header_res)
+                else:
+                    converted_data = pd.DataFrame(self.decode_binary(binary,time_patch,verbose)[:, [0,1]], columns=['tof', 'sweep'])     # saves only tof and sweep info
+                    converted_data.tof = converted_data.tof/10  # 100ps -> ns
+                    converted_data.sweep = converted_data.sweep.astype('int64')  # sweep is only int
+                    if to_csv:
+                        converted_data.to_csv('{}/{}.csv'.format(os.path.split(filename)[0],os.path.splitext(os.path.basename(filename))[0]), index=False)
+                print('File {} loaded successfully!'.format(os.path.splitext(os.path.basename(filename))[0]))
+                self.df_dict[os.path.splitext(os.path.basename(filename))[0]] = converted_data
+        if full_info == False:
+            return(pd.concat(self.df_dict, axis=1))    # convert dict of dataframes into one dataframe with two column name levels
+
+class MPANTMpa():
+    """
+    Original from https://gitlab.cern.ch/datanaso/dEval/-/blob/master/etc/mpant.py
+    Class handling the MPANT (mpa) data files!
+    Data format is 'asc'
+    """
+
+    def __init__(self):
+        self.data = {}
+        self.pars = {}
+        self.df_dict = {}
+        self.step = 0
+        # self.map_files = file_array
+
+    def increment(self):
+        self.step += 1
+
+    def read(self, mpa):
+        """
+        Read mpa data file
+        :return:
+        """
+        with open(mpa, 'r') as f:
+            fs = f.read()
+            name = os.path.basename(mpa).split('.')[0]
+            raw_header, raw_data = fs.split('[DATA]\n')
+            if bool(raw_data) and len(raw_data.split(' ')) >= 9:
+                self.parse_header(name, raw_header)
+                self.df_dict[name] = pd.read_csv(StringIO(raw_data), delimiter=' ', usecols=(0, 1, 2), header=None, names=['tof', 'sweep', 'counts'])
+                self.data[name] = self.df_dict[name].to_numpy()
+
+    def parse_header(self, key, txt):
+        parser = CustomParser(strict=False)
+        parser.read_file(StringIO(txt))
+        tmp = parser.as_dict()
+        self.pars[key] = dict(**tmp['CHN1'])
+        self.pars[key].update(**tmp['MPA4A'])
+
+    def process(self, files, to_csv = False):
+        for i, f in enumerate(files):
+            self.read(f)
+            # Convert bins to tof in ns
+            name = os.path.basename(f).split('.')[0]
+            self.df_dict[name].tof = self.df_dict[name].tof * float(self.pars[name]['calfact']) + float(self.pars[name]['caloff'])
+            if to_csv:
+                self.df_dict[name].to_csv('{}/{}.csv'.format(os.path.split(f)[0],os.path.splitext(os.path.basename(f))[0]), index=False)
+        #
+        return self.pars, self.data, pd.concat(self.df_dict, axis=1)
+
+class MCDWIN887():
+    """
+    Original from: https://gitlab.cern.ch/datanaso/dEval/-/blob/master/etc/mcdwin.py
+    Class handling the MCDWIN (.887) data files!
+    Pass the .887 file and the data format is 'csv, asc' is handled automatically.
+    """
+    def __init__(self):
+        self.data = {}
+        self.pars = {}
+        self.df_dict = {}
+        self.step = 0
+
+    def get_signals(self):
+        return self.next_887, self.status_887
+
+    def increment(self):
+        self.step += 1
+
+    def read(self, p887):
+        """
+        Read mpa data file
+        :return:
+        """
+        self.parse_header(p887)
+        folder = os.path.dirname(p887) + os.path.sep
+        name = os.path.basename(p887).split('.')[0]
+
+        if self.pars[name]['sweepmode'] == '80' and self.pars[name]['fmt'] == 'csv':
+            self.df_dict[name] = self.read_csv_80(folder, name)
+        elif self.pars[name]['sweepmode'] == '84' and self.pars[name]['fmt'] == 'csv':
+            self.df_dict[name] = self.read_csv_84(folder, name)
+        elif self.pars[name]['sweepmode'] == '84' and self.pars[name]['fmt'] == 'asc':
+            self.df_dict[name] = self.read_asc_84(folder, name)
+
+        if self.df_dict[name]['counts'].sum() != 0:
+            self.data[name] = self.df_dict[name][self.df_dict[name]['counts'] > 0].to_numpy()
+        else:
+            self.data[name] = self.df_dict[name][0:1].to_numpy()
+
+    def read_csv_80(self, folder, name):
+        self.pars[name]['cycles'] = '1'
+        fname = folder + name + f'.{self.pars[name]["fmt"]}'
+        df = pd.read_csv(fname, sep='\t', header=None, dtype=float, prefix='Y')
+
+        chan = [y for y in range(int(self.pars[name]['range']))]
+        cycles = [np.full((int(self.pars[name]['range'])), i) for i in range(1)]
+        slices = np.concatenate(cycles, axis=0)
+        return pd.DataFrame({'tof': np.array(chan), 'sweep': slices, 'counts': df['Y1'].to_numpy()})
+
+    def read_csv_84(self, folder, name):
+        fname = folder + name + f'.{self.pars[name]["fmt"]}'
+        df = pd.read_csv(fname, sep='\t', header=None, dtype=float, prefix='Y')
+
+        chan = [y for x in range(int(self.pars[name]['cycles'])) for y in range(int(self.pars[name]['range']))]
+        cycles = [np.full((int(self.pars[name]['range'])), i) for i in range(int(self.pars[name]['cycles']))]
+        slices = np.concatenate(cycles, axis=0)
+
+        return pd.DataFrame({'tof': np.array(chan), 'sweep': slices, 'counts': df['Y1'].to_numpy()})
+
+    def read_asc_84(self, folder, name):
+        fname = folder + name + f'.{self.pars[name]["fmt"]}'
+        with open(fname, 'r') as f:
+            fs = f.read()
+            _, raw_data = fs.split('[DATA]\n')
+            if len(raw_data.split(' ')) >= 3:  # If there is data (more than three lines), load it to df
+                df = pd.read_csv(StringIO(raw_data), delimiter=' ', usecols=(0, 1, 2), header=None, names=['tof', 'sweep', 'counts'])
+            else:  # Else build an empty dummy df
+                df = pd.DataFrame({'tof': np.zeros(5),
+                                   'sweep': np.zeros(5),
+                                   'counts': np.zeros(5)})
+        return df
+
+    def parse_header(self, key):
+        with open(key, 'r') as f:
+            fs = '[root]\n' + f.read()
+            key = os.path.basename(key).split('.')[0]
+            parser = CustomParser(strict=False)
+            parser.read_file(StringIO(fs))
+            tmp = parser.as_dict()
+
+            self.pars[key] = dict(**tmp['root'])
+            # correct for inconsistent 'fmt' keyword in FASTCom configuration file (.887)
+            if self.pars[key]['fmt'] == '3':
+                self.pars[key]['fmt_idx'] = '3'
+                self.pars[key]['fmt'] = 'asc'
+
+    def process(self, files, to_csv=True):
+        for i, f in enumerate(files):
+            self.read(f)
+            # Convert bins to tof in ns
+            name = os.path.basename(f).split('.')[0]
+            self.df_dict[name].tof = self.df_dict[name].tof * float(self.pars[name]['calfact']) + float(self.pars[name]['caloff'])
+            if to_csv:
+                self.df_dict[name].to_csv('{}/{}.csv'.format(os.path.split(f)[0],os.path.splitext(os.path.basename(f))[0]), index=False)
+
+        #
+        return self.pars, self.data, pd.concat(self.df_dict, axis=1)
